requirements.txt
streamlit                                   # UI
google-cloud-vision                         # Image unnderstanding / OCR - Optical Character Recognition
google-cloud-speech                         # Google Speech-to-Text client
google-cloud-storage                        # Load metadata & imahes from GCS
google-cloud-aiplatform                     # Gemini API & embeddings
vertexai                                    # Gemini API & embeddings
elevenlabs                                  # ElevenLabs API client
pydub                                       # Audio handling 
numpy                                       # Data processing library
pandas                                      # Data processing library
Pillow                                      # Image manipulation
python-dotenv                               # Load API keys 
tiktoken                                    # Token counting & safety